# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the CIJ project",
        "title": "CIJ Project API",
        "contact": {
            "name": "API Support",
            "email": "cauakathdev@gmail.com"
        },
        "version": "1.0"
    },
    "host": "conexao-inclusao.com",
    "basePath": "/",
    "paths": {
...
        "/people": {
            "get": {
                "description": "list all registered people and their users.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "List all registered people.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.PersonResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new person and their user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Create a new person.",
                "parameters": [
                    {
                        "description": "Person",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PersonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            }
        },
        "/people/:id": {
            "get": {
                "description": "get a person by their ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Get a person by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PersonResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "update an existent person and their user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Update a person.",
                "parameters": [
                    {
                        "description": "Person",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PersonRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete an existent person and their user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Delete a person.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            }
        },
        "/people/:id/address": {
            "put": {
                "description": "update an existent person address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Update a person address.",
                "parameters": [
                    {
                        "description": "Address",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddressRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            }
        },
        "/people/:id/disabilities": {
            "put": {
                "description": "update an existent person disabilities.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Update a person disabilities.",
                "parameters": [
                    {
                        "description": "Disabilities",
                        "name": "disabilities",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DisabilityRequest"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "enum.ColorBlindnessEnum": {
            "type": "string",
            "enum": [
                "normal",
                "protanopia",
                "deuteranopia",
                "tritanopia"
            ],
            "x-enum-varnames": [
                "Normal",
                "Protanopia",
                "Deuteranopia",
                "Tritanopia"
            ]
        },
        "enum.GenderEnum": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "other"
            ],
            "x-enum-varnames": [
                "Male",
                "Female",
                "Other"
            ]
        },
        "enum.ThemeEnum": {
            "type": "string",
            "enum": [
                "light",
                "dark",
                "system"
            ],
            "x-enum-varnames": [
                "Light",
                "Dark",
                "System"
            ]
        },
        "model.AddressRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
        "model.AddressResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "neighborhood": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
...
        "model.Config": {
            "type": "object",
            "properties": {
                "color_blindness": {
                    "$ref": "#/definitions/enum.ColorBlindnessEnum"
                },
                "font_size": {
                    "type": "integer"
                },
                "screen_reader": {
                    "type": "boolean"
                },
                "system_colors": {
                    "$ref": "#/definitions/model.SystemColors"
                },
                "theme": {
                    "$ref": "#/definitions/enum.ThemeEnum"
                },
                "voice_capture": {
                    "type": "boolean"
                }
            }
        },
...
        "model.DisabilityRequest": {
            "type": "object",
            "properties": {
                "acquired": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.Field": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.PersonDisabilityResponse": {
            "type": "object",
            "properties": {
                "acquired": {
                    "type": "boolean"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rate": {
                    "type": "string"
                }
            }
        },
        "model.PersonRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/model.AddressRequest"
                },
                "cpf": {
                    "type": "string"
                },
                "disabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DisabilityRequest"
                    }
                },
                "gender": {
                    "$ref": "#/definitions/enum.GenderEnum"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.UserRequest"
                }
            }
        },
        "model.PersonResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/model.AddressResponse"
                },
                "cpf": {
                    "type": "string"
                },
                "disabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PersonDisabilityResponse"
                    }
                },
                "gender": {
                    "$ref": "#/definitions/enum.GenderEnum"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                }
            }
        },
...
        "model.SystemChartColors": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "model.SystemColors": {
            "type": "object",
            "properties": {
                "chart_colors": {
                    "$ref": "#/definitions/model.SystemChartColors"
                },
                "primary_colors": {
                    "$ref": "#/definitions/model.SystemPrimaryColors"
                }
            }
        },
        "model.SystemPrimaryColors": {
            "type": "object",
            "properties": {
                "background_color": {
                    "type": "string"
                },
                "font_color": {
                    "type": "string"
                },
                "input_color": {
                    "type": "string"
                },
                "primary_color": {
                    "type": "string"
                },
                "secondary_color": {
                    "type": "string"
                },
                "secondary_font_color": {
                    "type": "string"
                }
            }
        },
        "model.UserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/model.Config"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "utils.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Field"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}